+++ c3IDE/Templates/ITemplate.cs
@@ -1,4 +1,6 @@
-﻿namespace c3IDE.Templates
+﻿using System;
+
+namespace c3IDE.Templates
 {
     public interface ITemplate
     {
@@ -18,5 +20,8 @@
 
         string EffectCode { get; }
         string EffectLangauge { get; }
+
+        string ThemeCode { get; }
+        string ThemeLanguage { get; }
     }
 }